Added name customisation option
This took a while because I needed to make sure that all the places where the names are mentioned are updated. I also made the form collapsible using Bootstrap and made it hide after the form is submitted. I had to move the serve tracker feature into a function to make sure that it updates the name whilst retaining the correct serve count when names are edited.