Add age and dad joke
Age uses random number 1-10 as foxes don't live very long. It looked weird when age is 0, so I made it so it changes to 6 months if it's 0.
Used the dad joke API so jokes are in English - in future I'd like to make them occasionally appear in Spanish though, so left that function in